Add PHPUnit test suite (unit + integration) and fix float comparison bug
Tests added:
- SanitizePriorityValueTest: valid priorities pass through; invalid/empty/SQL
  injection/mixed-case all return 'none' (5 assertions)
- AddCompletionColumnTest: column added with correct key and label (4 assertions)
- MakeCompletionSortableTest: sortable key registered, existing columns preserved (3)
- GetDueDateDisplayTest: all 4 due-date states (overdue/today/soon/later) with
  correct CSS class and label text (9 assertions)
- SaveCompletionStatusTest: nonce missing, nonce invalid, autosave, and
  capability guards all prevent update_post_meta calls (unit, 4 tests)
- SaveMetaIntegrationTest: real DB — saves yes/no/absent completion, valid/
  malformed/empty due date, valid/invalid priority, security guard checks (10)
- FilterPostsByCompletionTest: complete/incomplete/priority query filters,
  no-filter no-op (4 integration tests)

Bug fixed (found by tests):
- getDueDateDisplay() used `=== 0` (strict integer) on the return value of
  floor() which returns float. Posts due today were misclassified as 'due soon'.
  Fix: cast to int before comparison.

Infrastructure:
- composer.json with phpunit/phpunit ^9.6 and 10up/wp_mock ^1.1
- phpunit.xml: separate 'unit' and 'integration' test suites
- tests/bootstrap.php: WP_Mock + WordPress function stubs for unit suite
- tests/integration-bootstrap.php: real WP test library (SVN, Docker container)
- tests/wp-tests-config.php: Docker DB credentials for integration suite

Run unit tests:    vendor/bin/phpunit --testsuite unit
Run integration:   docker exec wp-wordpress-1 bash -c "cd /var/www/html/wp-content/plugins/draft-status && vendor/bin/phpunit --bootstrap tests/integration-bootstrap.php --testsuite integration"

+++ draft-status.php
@@ -151,7 +151,7 @@ class DraftStatus {
     private function getDueDateDisplay($due_date) {
         $due_timestamp = strtotime($due_date);
         $today = strtotime('today');
-        $days_diff = floor(($due_timestamp - $today) / (60 * 60 * 24));
+        $days_diff = (int) floor(($due_timestamp - $today) / (60 * 60 * 24));
 
         $date_class = 'draft-due-date';
         $date_label = '';
